+++ packages/ui-chat/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sharelyai/ui-chat
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @sharelyai/services@0.0.23
+  - @sharelyai/ui-shared@0.0.23
+
 ## 0.0.22
 
 ### Patch Changes